requirements.txt
Django==1.10.2
django-admin-tools==0.8.0

django-tastypie==0.13.3
factory_boy==2.7.0
selenium==3.0.0b3
